feat: invitationCode -> tripId & fetch 조인 사용

+++ src/main/java/com/ll/trip/domain/trip/trip/controller/TripController.java
@@ -15,11 +15,9 @@ import org.springframework.web.bind.annotation.RestController;
 import com.ll.trip.domain.file.file.service.AwsAuthService;
 import com.ll.trip.domain.trip.trip.dto.TripCreateDto;
 import com.ll.trip.domain.trip.trip.dto.TripInfoDto;
-import com.ll.trip.domain.trip.trip.dto.TripMemberDto;
 import com.ll.trip.domain.trip.trip.entity.Trip;
 import com.ll.trip.domain.trip.trip.service.TripService;
 import com.ll.trip.domain.user.user.entity.UserEntity;
-import com.ll.trip.domain.user.user.service.UserService;
 import com.ll.trip.global.security.userDetail.SecurityUser;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -30,7 +28,7 @@ import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.constraints.NotBlank;
+import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -42,8 +40,8 @@ import lombok.extern.slf4j.Slf4j;
 public class TripController {
 
 	private final TripService tripService;
-	private final UserService userService;
 	private final AwsAuthService awsAuthService;
+	private final EntityManager entityManager;
 
 	@PostMapping("/create")
 	@Operation(summary = "여행방 생성")
@@ -58,7 +56,7 @@ public class TripController {
 	) {
 		String invitationCode = tripService.generateInvitationCode();
 		Trip trip = tripService.createTrip(tripCreateDto, invitationCode);
-		UserEntity user = userService.findUserByUserId(securityUser.getId());
+		UserEntity user = entityManager.getReference(UserEntity.class, securityUser.getId());
 
 		tripService.joinTripById(trip, user, true);
 
@@ -69,18 +67,17 @@ public class TripController {
 	@Operation(summary = "여행방 참가")
 	@ApiResponse(responseCode = "200", description = "여행방 참가", content = {
 		@Content(mediaType = "application/json", schema = @Schema(implementation = TripInfoDto.class))})
-	public ResponseEntity<?> joinTripByInvitationCode(
+	public ResponseEntity<?> joinTrip(
 		@AuthenticationPrincipal SecurityUser securityUser,
-		@RequestParam @NotBlank @Parameter(description = "초대 코드", example = "1A2B3C4D") String invitationCode
+		@RequestParam @Parameter(description = "초대코드", example = "1A2B3C4D") String invitationCode
 	) {
-		Trip trip = tripService.findByInvitationCode(invitationCode);
-		UserEntity user = userService.findUserByUserId(securityUser.getId());
+		long tripId = tripService.findTripIdByInvitationCode(invitationCode);
+		Trip trip = entityManager.getReference(Trip.class, tripId);
+		UserEntity user = entityManager.getReference(UserEntity.class, securityUser.getId());
 
 		tripService.joinTripById(trip, user, false);
 
-		TripInfoDto response = new TripInfoDto(trip);
-		List<TripMemberDto> tripMemberDtoList = tripService.findTripMemberUserByTripId(trip.getId());
-		response.setTripMemberDtoList(tripMemberDtoList);
+		TripInfoDto response = new TripInfoDto(tripService.findTripDetailByTripId(tripId));
 
 		return ResponseEntity.ok(response);
 	}
@@ -89,17 +86,14 @@ public class TripController {
 	@Operation(summary = "여행방 입장")
 	@ApiResponse(responseCode = "200", description = "여행방 입장", content = {
 		@Content(mediaType = "application/json", schema = @Schema(implementation = TripInfoDto.class))})
-	public ResponseEntity<?> enterTripByInvitationCode(
+	public ResponseEntity<?> enterTrip(
 		@AuthenticationPrincipal SecurityUser securityUser,
-		@RequestParam @NotBlank @Parameter(description = "초대 코드", example = "1A2B3C4D") String invitationCode
+		@RequestParam @Parameter(description = "트립 id", example = "1") long tripId
 	) {
-		Trip trip = tripService.findByInvitationCode(invitationCode);
-		if (!tripService.existTripMemberByTripIdAndUserId(trip.getId(), securityUser.getId()))
+		if (!tripService.existTripMemberByTripIdAndUserId(tripId, securityUser.getId()))
 			return ResponseEntity.badRequest().body("입장 권한이 없습니다.");
 
-		TripInfoDto response = new TripInfoDto(trip);
-		List<TripMemberDto> tripMemberDtoList = tripService.findTripMemberUserByTripId(trip.getId());
-		response.setTripMemberDtoList(tripMemberDtoList);
+		TripInfoDto response = new TripInfoDto(tripService.findTripDetailByTripId(tripId));
 
 		return ResponseEntity.ok(response);
 	}
@@ -108,21 +102,11 @@ public class TripController {
 	@Operation(summary = "다가오는 Trip리스트 요청")
 	@ApiResponse(responseCode = "200", description = "다가오는 Trip리스트 요청", content = {
 		@Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TripInfoDto.class)))})
-	public ResponseEntity<?> showIncomingTripListByUserId(
-		@AuthenticationPrincipal SecurityUser securityUser,
-		@RequestParam(required = false) @Parameter(description = "정렬기준 필드", example = "startDate : 시작 날짜 순, endDate : 끝나는 날짜 순") String sortField,
-		@RequestParam(required = false) @Parameter(description = "정렬순서", example = "ASC : 오름차순&오래된순, DESC : 내림차순&최신순") String sortDirection
+	public ResponseEntity<?> showIncomingTripList(
+		@AuthenticationPrincipal SecurityUser securityUser
 	) {
-		//TODO 북마크, 다가오는, 지난 + queryDSL로 동적 쿼리 생성
-		List<TripInfoDto> response = null;
-		if (sortDirection == null && sortField == null) {
-			response = tripService.findAllByUserId(securityUser.getId(), LocalDate.now(), "incoming",
-				"startDate", "DESC");
-		} else if (sortDirection != null && sortField != null) {
-			tripService.findAllByUserId(securityUser.getId(), LocalDate.now(), "incoming",
-				sortDirection, sortField);
-		}
-		tripService.fillTripMemberToTripInfo(response);
+		List<TripInfoDto> response = tripService.findAllIncomingByUserId(securityUser.getId(), LocalDate.now());
+
 		return ResponseEntity.ok(response);
 	}
 
@@ -130,54 +114,40 @@ public class TripController {
 	@Operation(summary = "지난 Trip리스트 요청")
 	@ApiResponse(responseCode = "200", description = "지난 Trip리스트 요청", content = {
 		@Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TripInfoDto.class)))})
-	public ResponseEntity<?> showLastTripListByUserId(
-		@AuthenticationPrincipal SecurityUser securityUser,
-		@RequestParam @Parameter(description = "정렬기준 필드", example = "startDate : 시작 날짜 순, endDate : 끝나는 날짜 순") String sortField,
-		@RequestParam @Parameter(description = "정렬순서", example = "ASC : 오름차순&오래된순, DESC : 내림차순&최신순") String sortDirection
+	public ResponseEntity<?> showLastTripList(
+		@AuthenticationPrincipal SecurityUser securityUser
 	) {
-		//TODO 북마크, 다가오는, 지난 + queryDSL로 동적 쿼리 생성
-		List<TripInfoDto> response = null;
-		if (sortDirection == null && sortField == null) {
-			response = tripService.findAllByUserId(securityUser.getId(), LocalDate.now(), "last",
-				"startDate", "DESC");
-		} else if (sortDirection != null && sortField != null) {
-			tripService.findAllByUserId(securityUser.getId(), LocalDate.now(), "last",
-				sortDirection, sortField);
-		}
-		tripService.fillTripMemberToTripInfo(response);
+		List<TripInfoDto> response = tripService.findAllLastByUserIdAndDate(securityUser.getId(), LocalDate.now());
 		return ResponseEntity.ok(response);
 	}
 
 	@GetMapping("/list/bookmark")
-	@Operation(summary = "지난 Trip리스트 요청")
-	@ApiResponse(responseCode = "200", description = "지난 Trip리스트 요청", content = {
+	@Operation(summary = "Trip 북마크 리스트 요청")
+	@ApiResponse(responseCode = "200", description = "Trip 북마크 리스트 요청", content = {
 		@Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TripInfoDto.class)))})
-	public ResponseEntity<?> showBookmarkTripListByUserId(
-		@AuthenticationPrincipal SecurityUser securityUser,
-		@RequestParam @Parameter(description = "정렬기준 필드", example = "startDate : 시작 날짜 순, endDate : 끝나는 날짜 순") String sortField,
-		@RequestParam @Parameter(description = "정렬순서", example = "ASC : 오름차순&오래된순, DESC : 내림차순&최신순") String sortDirection
+	public ResponseEntity<?> showBookmarkTripList(
+		@AuthenticationPrincipal SecurityUser securityUser
 	) {
-		//TODO 북마크, 다가오는, 지난 + queryDSL로 동적 쿼리 생성
-		List<TripInfoDto> response = tripService.findBookmarkByUserId(securityUser.getId(), sortDirection, sortField);
-		tripService.fillTripMemberToTripInfo(response);
+		List<TripInfoDto> response = tripService.findBookmarkByUserId(securityUser.getId());
 		return ResponseEntity.ok(response);
 	}
 
 	@GetMapping("/delete")
 	@Operation(summary = "Trip 삭제")
 	@ApiResponse(responseCode = "200", description = "Trip 삭제")
-	public ResponseEntity<?> deleteTripByInvitationCode(
+	public ResponseEntity<?> deleteTrip(
 		@AuthenticationPrincipal SecurityUser securityUser,
-		@RequestParam @Parameter(description = "초대코드", example = "1A2B3C4D") String invitationCode
+		@RequestParam @Parameter(description = "트립 id", example = "1") long tripId
 	) {
-		Trip trip = tripService.findByInvitationCode(invitationCode);
+		if (!tripService.isLeaderOfTrip(securityUser.getId(), tripId))
+			return ResponseEntity.badRequest().body("해당 여행방에 대한 수정/삭제 권한이 없습니다.");
 
-		List<String> urls = awsAuthService.abstractUrlFromPresignedUrl(List.of(trip.getThumbnail()));
-		List<String> keys = awsAuthService.abstractKeyFromUrl(urls);
+		List<String> urls = tripService.findImageByTripId(tripId);
+		List<String> keys = awsAuthService.extractKeyFromUrl(urls);
 
 		awsAuthService.deleteObjectByKey(keys);
 
-		tripService.deleteTripById(trip.getId());
+		tripService.deleteTripById(tripId);
 
 		return ResponseEntity.ok("deleted");
 	}
@@ -186,13 +156,14 @@ public class TripController {
 	@Operation(summary = "Trip 수정")
 	@ApiResponse(responseCode = "200", description = "Trip 수정", content = {
 		@Content(mediaType = "application/json", schema = @Schema(implementation = TripInfoDto.class))})
-	public ResponseEntity<?> modifyTripByInvitationCode(
+	public ResponseEntity<?> modifyTrip(
 		@AuthenticationPrincipal SecurityUser securityUser,
+		@RequestParam @Parameter(description = "트립 id", example = "1") long tripId,
 		@RequestBody TripInfoDto requestBody
 	) {
-		Trip trip = tripService.findByInvitationCode(requestBody.getInvitationCode());
-
-		TripInfoDto response = tripService.modifyTripByDto(trip, requestBody);
+		if (!tripService.isLeaderOfTrip(securityUser.getId(), tripId))
+			return ResponseEntity.badRequest().body("해당 여행방에 대한 수정/삭제 권한이 없습니다.");
+		TripInfoDto response = tripService.modifyTripByDto(entityManager.getReference(Trip.class, tripId), requestBody);
 
 		return ResponseEntity.ok(response);
 	}
@@ -202,14 +173,17 @@ public class TripController {
 	@Operation(summary = "Trip 북마크 토글")
 	@ApiResponse(responseCode = "200", description = "Trip 북마크 토글", content = {
 		@Content(mediaType = "application/json", schema = @Schema(implementation = Boolean.class))})
-	public ResponseEntity<?> toggleTripBookmarkByInvitationCode(
+	public ResponseEntity<?> toggleTripBookmark(
 		@AuthenticationPrincipal SecurityUser securityUser,
-		@RequestParam @Parameter(description = "초대코드", example = "1A2B3C4D") String invitationCode
+		@RequestParam @Parameter(description = "트립 id", example = "1") long tripId
 	) {
-		Trip trip = tripService.findByInvitationCode(invitationCode);
-		//날짜 수정시 기존 플랜변경은 미정
-		boolean response = tripService.toggleBookmarkByTripAndUserId(trip, securityUser.getId());
+		int update = tripService.toggleBookmarkByTripAndUserId(securityUser.getId(), tripId);
+		if (update == 0)
+			tripService.createTripBookmark(
+				entityManager.getReference(UserEntity.class, securityUser.getId()),
+				entityManager.getReference(Trip.class, tripId)
+			);
 
-		return ResponseEntity.ok(response);
+		return ResponseEntity.ok(tripService.getIsToggleByUserIdAndScrapId(securityUser.getId(), tripId));
 	}
 }
